refactor(core): extract DeploymentUrlBuilder into rise-backend-core
Split the backend-agnostic URL/image resolution out of the K8s
`ResourceBuilder` into a new `DeploymentUrlBuilder` in `rise-backend-core`. It
holds the ingress URL templates + registry provider and owns
`compute_deployment_urls`/`compute_project_urls`/`resolve_image`/
`primary_ingress_hosts` and their helpers (`IngressUrl`, the `resolved_*`/
`full_ingress_url*`/`host_to_url` family, `filter_valid_custom_domains`). The
custom-domain conflict validator moves to `core::custom_domain`, and the
in-container identity-mount constants to `core::identity`; both are re-exported
from their old paths.

The K8s `ResourceBuilder` now composes a `DeploymentUrlBuilder` and delegates
its URL/image methods to it, so K8s behavior is unchanged. The Docker backend
and reconciler hold an `Arc<DeploymentUrlBuilder>` directly instead of an
`Arc<ResourceBuilder>` — which removes the last `k8s_openapi` reference from the
Docker controller. `init_docker_backend` now builds a `DeploymentUrlBuilder`
rather than a full K8s `ResourceBuilder`.

Behavior-preserving. Verified: CLI + backend builds, clippy `--all-targets`
clean; 47 core, 27 webhook, and 163 resource_builder/Docker controller tests
pass; the Docker controller no longer references `k8s_openapi`.
