+++ src/server/deployment/resource_builder.rs
@@ -22,15 +22,16 @@ use k8s_openapi::api::networking::v1::{
 use k8s_openapi::apimachinery::pkg::apis::meta::v1::{LabelSelector, ObjectMeta};
 use k8s_openapi::ByteString;
 use std::collections::BTreeMap;
-use std::sync::Arc;
 use tracing::warn;
 
 use crate::db::models::{CustomDomain, Deployment, Project};
-use crate::server::registry::{
-    models::{RegistryAuthMethod, RegistryCredentials},
-    RegistryProvider,
-};
+use crate::server::registry::models::{RegistryAuthMethod, RegistryCredentials};
 use crate::server::settings::AccessRequirement;
+use rise_backend_core::DeploymentUrlBuilder;
+// The parsed-ingress-URL type and the URL/image computation live in
+// `rise-backend-core`; re-exported so existing `resource_builder::IngressUrl`
+// consumers (e.g. the Docker reconciler) keep their import path.
+pub use rise_backend_core::IngressUrl;
 
 // Re-export constants used by webhook and other consumers
 pub const LABEL_MANAGED_BY: &str = "app.kubernetes.io/managed-by";
@@ -49,17 +50,12 @@ const EXTRA_SERVICE_TOKENS_MOUNT_PATH: &str = "/var/run/secrets/rise/tokens";
 // Workload identity: a per-deployment Secret carrying the bootstrap credential
 // and any auto-minted workload tokens, mounted as files into the app's pod.
 const IDENTITY_VOLUME_NAME: &str = "rise-identity";
-/// In-container mount path for the workload-identity material. This is part of
-/// the backend-agnostic contract a workload reads, so the Docker backend (which
-/// delivers the same files via the Docker archive API rather than a Secret
-/// volume) reuses this constant to land the files at the same path.
-pub const IDENTITY_MOUNT_PATH: &str = "/var/run/secrets/rise/identity";
-/// Subdirectory under [`IDENTITY_MOUNT_PATH`] holding the auto-minted per-audience
-/// token files. Shared with the Docker backend so both backends deliver tokens to
-/// the same in-container path.
-pub const IDENTITY_TOKENS_SUBDIR: &str = "tokens";
-/// Secret data key for the bootstrap credential.
-pub const IDENTITY_CREDENTIAL_KEY: &str = "credential";
+// The in-container identity contract (mount path, tokens subdir, credential key)
+// is backend-agnostic and lives in `rise-backend-core`; re-exported here for
+// the K8s builder and its existing consumers.
+pub use rise_backend_core::identity::{
+    IDENTITY_CREDENTIAL_KEY, IDENTITY_MOUNT_PATH, IDENTITY_TOKENS_SUBDIR,
+};
 /// Prefix for Secret data keys holding auto-minted workload tokens.
 pub const IDENTITY_TOKEN_KEY_PREFIX: &str = "token-";
 
@@ -126,22 +122,12 @@ pub struct ContainerRuntime<'a> {
     pub health_check: Option<crate::server::deployment::models::HealthCheckSpec>,
 }
 
-/// Parsed ingress URL components
-#[derive(Debug, Clone)]
-pub struct IngressUrl {
-    pub host: String,
-    pub path_prefix: Option<String>,
-}
-
 /// Holds configuration for building Kubernetes resource specs.
 /// Does NOT hold a kube client — pure spec generation only.
 pub struct ResourceBuilder {
-    pub production_ingress_url_template: String,
-    pub staging_ingress_url_template: Option<String>,
-    pub environment_ingress_url_template: Option<String>,
-    pub ingress_port: Option<u16>,
-    pub ingress_schema: String,
-    pub registry_provider: Arc<dyn RegistryProvider>,
+    /// Backend-agnostic URL/image resolution (ingress templates + registry).
+    /// The K8s-specific methods below delegate their URL/image work to it.
+    pub url_builder: DeploymentUrlBuilder,
     pub auth_backend_url: String,
     pub auth_signin_url: String,
     pub backend_address: Option<crate::server::settings::BackendAddress>,
@@ -175,15 +161,11 @@ impl ResourceBuilder {
     }
 
     pub fn sanitize_label_value(value: &str) -> String {
-        crate::server::deployment::models::normalize_deployment_group(value)
+        DeploymentUrlBuilder::sanitize_label_value(value)
     }
 
     pub fn escaped_group_name(deployment_group: &str) -> String {
-        if deployment_group == crate::server::deployment::models::DEFAULT_DEPLOYMENT_GROUP {
-            "default".to_string()
-        } else {
-            Self::sanitize_label_value(deployment_group)
-        }
+        DeploymentUrlBuilder::escaped_group_name(deployment_group)
     }
 
     pub fn service_name(_project: &Project, deployment: &Deployment) -> String {
@@ -221,154 +203,20 @@ impl ResourceBuilder {
         format!("env-{}", environment_name)
     }
 
-    // ── URL resolution ─────────────────────────────────────────────────
+    // ── URL resolution + deployment URLs ───────────────────────────────
+    //
+    // Backend-agnostic; the implementations live on the composed
+    // `DeploymentUrlBuilder`. Thin wrappers are kept for the methods the K8s
+    // resource builders and the HTTP handlers still call.
 
-    fn parse_ingress_url(url: &str) -> IngressUrl {
-        match url.find('/') {
-            Some(slash_pos) => IngressUrl {
-                host: url[..slash_pos].to_string(),
-                path_prefix: Some(url[slash_pos..].to_string()),
-            },
-            None => IngressUrl {
-                host: url.to_string(),
-                path_prefix: None,
-            },
-        }
-    }
-
-    pub fn resolved_ingress_url(&self, project: &Project, deployment: &Deployment) -> String {
-        self.resolved_ingress_url_for_group(project, &deployment.deployment_group)
-    }
-
-    pub fn resolved_ingress_url_for_group(
+    pub(crate) fn ingress_url_components(
         &self,
         project: &Project,
-        deployment_group: &str,
-    ) -> String {
-        if deployment_group == crate::server::deployment::models::DEFAULT_DEPLOYMENT_GROUP {
-            self.production_ingress_url_template
-                .replace("{project_name}", &project.name)
-        } else if let Some(ref staging_template) = self.staging_ingress_url_template {
-            staging_template
-                .replace("{project_name}", &project.name)
-                .replace(
-                    "{deployment_group}",
-                    &Self::escaped_group_name(deployment_group),
-                )
-        } else {
-            let base_url = self
-                .production_ingress_url_template
-                .replace("{project_name}", &project.name);
-            if let Some(dot_pos) = base_url.find('.') {
-                format!(
-                    "{}-{}{}",
-                    &base_url[..dot_pos],
-                    Self::escaped_group_name(deployment_group),
-                    &base_url[dot_pos..]
-                )
-            } else {
-                format!(
-                    "{}-{}",
-                    base_url,
-                    Self::escaped_group_name(deployment_group)
-                )
-            }
-        }
+        deployment: &Deployment,
+    ) -> IngressUrl {
+        self.url_builder.ingress_url_components(project, deployment)
     }
 
-    /// Resolve the deployment-group-specific host (the staging-template form).
-    ///
-    /// Always returns the templated DG host when `staging_ingress_url_template` is set,
-    /// even for the "default" group (yielding e.g. `myapp-default.preview.example.com`).
-    /// When no staging template is configured, returns the synthesized form for
-    /// non-default groups only — for the default group it returns `None` to avoid
-    /// colliding with the project's production hostname.
-    pub fn resolved_deployment_group_url(
-        &self,
-        project: &Project,
-        deployment_group: &str,
-    ) -> Option<String> {
-        if let Some(ref staging_template) = self.staging_ingress_url_template {
-            Some(
-                staging_template
-                    .replace("{project_name}", &project.name)
-                    .replace(
-                        "{deployment_group}",
-                        &Self::escaped_group_name(deployment_group),
-                    ),
-            )
-        } else if deployment_group != crate::server::deployment::models::DEFAULT_DEPLOYMENT_GROUP {
-            let base_url = self
-                .production_ingress_url_template
-                .replace("{project_name}", &project.name);
-            Some(if let Some(dot_pos) = base_url.find('.') {
-                format!(
-                    "{}-{}{}",
-                    &base_url[..dot_pos],
-                    Self::escaped_group_name(deployment_group),
-                    &base_url[dot_pos..]
-                )
-            } else {
-                format!(
-                    "{}-{}",
-                    base_url,
-                    Self::escaped_group_name(deployment_group)
-                )
-            })
-        } else {
-            None
-        }
-    }
-
-    pub fn resolved_environment_url(
-        &self,
-        project: &Project,
-        environment: &crate::db::models::Environment,
-    ) -> Option<String> {
-        if environment.is_production {
-            Some(
-                self.production_ingress_url_template
-                    .replace("{project_name}", &project.name),
-            )
-        } else {
-            self.environment_ingress_url_template
-                .as_ref()
-                .map(|env_template| {
-                    env_template
-                        .replace("{project_name}", &project.name)
-                        .replace("{environment}", &environment.name)
-                })
-        }
-    }
-
-    fn full_ingress_url_from_host(&self, url: &str) -> String {
-        if let Some(port) = self.ingress_port {
-            let parsed = Self::parse_ingress_url(url);
-            let host_with_port = format!("{}:{}", parsed.host, port);
-            match parsed.path_prefix {
-                Some(path) => format!("{}{}", host_with_port, path),
-                None => host_with_port,
-            }
-        } else {
-            url.to_string()
-        }
-    }
-
-    fn full_ingress_url(&self, project: &Project, deployment: &Deployment) -> String {
-        let url = self.resolved_ingress_url(project, deployment);
-        self.full_ingress_url_from_host(&url)
-    }
-
-    fn ingress_url_components(&self, project: &Project, deployment: &Deployment) -> IngressUrl {
-        let url = self.resolved_ingress_url(project, deployment);
-        Self::parse_ingress_url(&url)
-    }
-
-    /// Returns `true` if `host` would also be emitted as the env URL of some
-    /// environment whose `primary_deployment_group` is *not* `deployment_group`.
-    /// That's the cross-ingress collision case: a deployment-group's name
-    /// happens to match an environment name, so its DG URL would steal a host
-    /// that "belongs" to the env's primary group.
     fn host_conflicts_with_other_env(
         &self,
         project: &Project,
@@ -376,31 +224,14 @@ impl ResourceBuilder {
         host: &str,
         all_environments: &[crate::db::models::Environment],
     ) -> bool {
-        all_environments.iter().any(|env| {
-            if env.primary_deployment_group.as_deref() == Some(deployment_group) {
-                return false;
-            }
-            match self.resolved_environment_url(project, env) {
-                Some(env_url) => Self::parse_ingress_url(&env_url).host == host,
-                None => false,
-            }
-        })
+        self.url_builder.host_conflicts_with_other_env(
+            project,
+            deployment_group,
+            host,
+            all_environments,
+        )
     }
 
-    /// Compute the ordered, deduplicated list of hosts that should appear on the
-    /// primary ingress for `(project, deployment_group)`.
-    ///
-    /// The list is built from:
-    /// 1. The deployment-group host (`resolved_deployment_group_url`), when one
-    ///    is configured for this group AND its host doesn't collide with the
-    ///    env URL of an environment whose primary group is a different group.
-    /// 2. The host of the environment whose `primary_deployment_group` matches
-    ///    this group, if any (production env contributes the production host).
-    /// 3. Each custom domain attached to that environment, when
-    ///    `include_custom_domains_inline` is true.
-    ///
-    /// `all_environments` is the full project env list used for the collision
-    /// check in (1). Callers without env context can pass an empty slice.
     pub fn primary_ingress_hosts(
         &self,
         project: &Project,
@@ -410,93 +241,21 @@ impl ResourceBuilder {
         include_custom_domains_inline: bool,
         all_environments: &[crate::db::models::Environment],
     ) -> Vec<IngressUrl> {
-        let mut hosts: Vec<IngressUrl> = Vec::new();
-        let mut seen: std::collections::HashSet<String> = std::collections::HashSet::new();
-
-        if let Some(dg_url) = self.resolved_deployment_group_url(project, deployment_group) {
-            let parsed = Self::parse_ingress_url(&dg_url);
-            if self.host_conflicts_with_other_env(
-                project,
-                deployment_group,
-                &parsed.host,
-                all_environments,
-            ) {
-                warn!(
-                    project = %project.name,
-                    deployment_group = %deployment_group,
-                    host = %parsed.host,
-                    "Suppressing deployment-group URL: collides with env URL of an environment whose primary deployment group is different"
-                );
-            } else if seen.insert(parsed.host.clone()) {
-                hosts.push(parsed);
-            }
-        }
-
-        if let Some(env) = environment_for_group {
-            if let Some(env_url) = self.resolved_environment_url(project, env) {
-                let parsed = Self::parse_ingress_url(&env_url);
-                if seen.insert(parsed.host.clone()) {
-                    hosts.push(parsed);
-                }
-            }
-        }
-
-        if include_custom_domains_inline {
-            for cd in custom_domains_for_env {
-                if seen.insert(cd.domain.clone()) {
-                    hosts.push(IngressUrl {
-                        host: cd.domain.clone(),
-                        path_prefix: None,
-                    });
-                }
-            }
-        }
-
-        hosts
+        self.url_builder.primary_ingress_hosts(
+            project,
+            deployment_group,
+            environment_for_group,
+            custom_domains_for_env,
+            include_custom_domains_inline,
+            all_environments,
+        )
     }
 
     pub fn filter_valid_custom_domains(
         &self,
         custom_domains: &[CustomDomain],
     ) -> Vec<CustomDomain> {
-        use crate::server::custom_domains::validation;
-
-        custom_domains
-            .iter()
-            .filter(|domain| {
-                match validation::validate_custom_domain(
-                    &domain.domain,
-                    &self.production_ingress_url_template,
-                    self.staging_ingress_url_template.as_deref(),
-                    self.environment_ingress_url_template.as_deref(),
-                    None,
-                ) {
-                    Ok(()) => true,
-                    Err(reason) => {
-                        warn!(
-                            domain_id = %domain.id,
-                            project_id = %domain.project_id,
-                            domain = %domain.domain,
-                            "Ignoring custom domain that conflicts with project default domain pattern: {}",
-                            reason
-                        );
-                        false
-                    }
-                }
-            })
-            .cloned()
-            .collect()
-    }
-
-    // ── Deployment URLs ────────────────────────────────────────────────
-
-    fn host_to_url(&self, host: &str) -> String {
-        let host_with_port = if let Some(port) = self.ingress_port {
-            format!("{}:{}", host, port)
-        } else {
-            host.to_string()
-        };
-        format!("{}://{}", self.ingress_schema, host_with_port)
+        self.url_builder.filter_valid_custom_domains(custom_domains)
     }
 
     pub fn compute_deployment_urls(
@@ -507,69 +266,13 @@ impl ResourceBuilder {
         all_environments: &[crate::db::models::Environment],
         custom_domains: &[CustomDomain],
     ) -> super::controller::DeploymentUrls {
-        let default_url_host = self.full_ingress_url(project, deployment);
-        let default_url = format!("{}://{}", self.ingress_schema, default_url_host);
-
-        let environment_for_group = environment.filter(|env| {
-            env.primary_deployment_group.as_deref() == Some(&deployment.deployment_group)
-        });
-
-        let environment_url = environment_for_group.and_then(|env| {
-            self.resolved_environment_url(project, env).map(|url_host| {
-                let full_host = self.full_ingress_url_from_host(&url_host);
-                format!("{}://{}", self.ingress_schema, full_host)
-            })
-        });
-
-        let valid_custom_domains = self.filter_valid_custom_domains(custom_domains);
-        let env_id = environment_for_group.map(|e| e.id);
-        let domains_for_env: Vec<&CustomDomain> = valid_custom_domains
-            .iter()
-            .filter(|cd| env_id == Some(cd.environment_id))
-            .collect();
-
-        let starred = domains_for_env.iter().find(|d| d.is_primary);
-        let primary_url = if let Some(starred) = starred {
-            self.host_to_url(&starred.domain)
-        } else {
-            environment_url
-                .clone()
-                .unwrap_or_else(|| default_url.clone())
-        };
-
-        let custom_domain_urls: Vec<String> = domains_for_env
-            .iter()
-            .map(|d| self.host_to_url(&d.domain))
-            .collect();
-
-        let templated_hosts = self.primary_ingress_hosts(
+        self.url_builder.compute_deployment_urls(
             project,
-            &deployment.deployment_group,
-            environment_for_group,
-            &[],
-            false,
+            deployment,
+            environment,
             all_environments,
-        );
-        let mut all_urls: Vec<String> = Vec::new();
-        let mut seen: std::collections::HashSet<String> = std::collections::HashSet::new();
-        for host in &templated_hosts {
-            let url = self.host_to_url(&host.host);
-            if seen.insert(url.clone()) {
-                all_urls.push(url);
-            }
-        }
-        for url in &custom_domain_urls {
-            if seen.insert(url.clone()) {
-                all_urls.push(url.clone());
-            }
-        }
-
-        super::controller::DeploymentUrls {
-            default_url,
-            primary_url,
-            custom_domain_urls,
-            all_urls,
-        }
+            custom_domains,
+        )
     }
 
     pub fn compute_project_urls(
@@ -578,48 +281,8 @@ impl ResourceBuilder {
         deployment_group: &str,
         custom_domains: &[CustomDomain],
     ) -> super::controller::DeploymentUrls {
-        let url_host = self.resolved_ingress_url_for_group(project, deployment_group);
-        let full_host = self.full_ingress_url_from_host(&url_host);
-        let default_url = format!("{}://{}", self.ingress_schema, full_host);
-
-        let valid_custom_domains = self.filter_valid_custom_domains(custom_domains);
-
-        // No environment context here — keep the historical "default group hosts custom
-        // domains" rule to preserve preview-URL semantics for callers that haven't yet
-        // adopted the env-aware path.
-        let (custom_domain_urls, primary_url) =
-            if deployment_group == crate::server::deployment::models::DEFAULT_DEPLOYMENT_GROUP {
-                let starred = valid_custom_domains.iter().find(|d| d.is_primary);
-                let primary = if let Some(starred) = starred {
-                    self.host_to_url(&starred.domain)
-                } else {
-                    default_url.clone()
-                };
-
-                let urls: Vec<String> = valid_custom_domains
-                    .iter()
-                    .map(|d| self.host_to_url(&d.domain))
-                    .collect();
-                (urls, primary)
-            } else {
-                (Vec::new(), default_url.clone())
-            };
-
-        let mut all_urls = vec![default_url.clone()];
-        let mut seen: std::collections::HashSet<String> = std::collections::HashSet::new();
-        seen.insert(default_url.clone());
-        for url in &custom_domain_urls {
-            if seen.insert(url.clone()) {
-                all_urls.push(url.clone());
-            }
-        }
-
-        super::controller::DeploymentUrls {
-            default_url,
-            primary_url,
-            custom_domain_urls,
-            all_urls,
-        }
+        self.url_builder
+            .compute_project_urls(project, deployment_group, custom_domains)
     }
 
     // ── Labels ─────────────────────────────────────────────────────────
@@ -1403,7 +1066,8 @@ impl ResourceBuilder {
                         image_pull_secrets: {
                             let secret_name =
                                 self.image_pull_secret_name.as_deref().or_else(|| {
-                                    self.registry_provider
+                                    self.url_builder
+                                        .registry_provider
                                         .requires_pull_secret()
                                         .then_some(IMAGE_PULL_SECRET_NAME)
                                 });
@@ -1568,7 +1232,7 @@ impl ResourceBuilder {
                 let signin_url = if self.backend_address.is_some() {
                     format!(
                         "{}://$http_host/.rise/auth/signin?project={}&redirect=$scheme://$http_host$escaped_request_uri",
-                        self.ingress_schema,
+                        self.url_builder.ingress_schema,
                         urlencoding::encode(&project.name)
                     )
                 } else {
@@ -1962,18 +1626,8 @@ impl ResourceBuilder {
         deployment: &Deployment,
         source_deployment_id: Option<&str>,
     ) -> String {
-        if let Some(ref image_digest) = deployment.image_digest {
-            image_digest.clone()
-        } else {
-            let deployment_id_for_tag = source_deployment_id
-                .unwrap_or(&deployment.deployment_id)
-                .to_string();
-            self.registry_provider.get_image_tag(
-                &project.name,
-                &deployment_id_for_tag,
-                crate::server::registry::ImageTagType::Internal,
-            )
-        }
+        self.url_builder
+            .resolve_image(project, deployment, source_deployment_id)
     }
 }
 
@@ -2032,12 +1686,14 @@ mod tests {
 
     fn test_resource_builder() -> ResourceBuilder {
         ResourceBuilder {
-            production_ingress_url_template: "{project_name}.example.test".to_string(),
-            staging_ingress_url_template: None,
-            environment_ingress_url_template: None,
-            ingress_port: None,
-            ingress_schema: "https".to_string(),
-            registry_provider: Arc::new(TestRegistryProvider::default()),
+            url_builder: DeploymentUrlBuilder {
+                production_ingress_url_template: "{project_name}.example.test".to_string(),
+                staging_ingress_url_template: None,
+                environment_ingress_url_template: None,
+                ingress_port: None,
+                ingress_schema: "https".to_string(),
+                registry_provider: Arc::new(TestRegistryProvider::default()),
+            },
             auth_backend_url: "https://auth.example.test".to_string(),
             auth_signin_url: "https://signin.example.test".to_string(),
             backend_address: None,
@@ -2408,7 +2064,7 @@ mod tests {
         requires_pull_secret: bool,
     ) -> K8sDeployment {
         let mut builder = test_resource_builder();
-        builder.registry_provider = Arc::new(TestRegistryProvider {
+        builder.url_builder.registry_provider = Arc::new(TestRegistryProvider {
             requires_pull_secret,
         });
         builder.image_pull_secret_name = image_pull_secret_name.map(str::to_string);
@@ -2517,7 +2173,9 @@ mod tests {
         let builder = test_resource_builder();
         let project = test_project();
         assert_eq!(
-            builder.resolved_deployment_group_url(&project, "default"),
+            builder
+                .url_builder
+                .resolved_deployment_group_url(&project, "default"),
             None
         );
     }
@@ -2527,7 +2185,9 @@ mod tests {
         let builder = test_resource_builder();
         let project = test_project();
         assert_eq!(
-            builder.resolved_deployment_group_url(&project, "mr-42"),
+            builder
+                .url_builder
+                .resolved_deployment_group_url(&project, "mr-42"),
             Some("demo-mr-42.example.test".to_string())
         );
     }
@@ -2535,15 +2195,19 @@ mod tests {
     #[test]
     fn dg_host_uses_staging_template_when_set() {
         let mut builder = test_resource_builder();
-        builder.staging_ingress_url_template =
+        builder.url_builder.staging_ingress_url_template =
             Some("{project_name}-{deployment_group}.preview.example.test".to_string());
         let project = test_project();
         assert_eq!(
-            builder.resolved_deployment_group_url(&project, "default"),
+            builder
+                .url_builder
+                .resolved_deployment_group_url(&project, "default"),
             Some("demo-default.preview.example.test".to_string())
         );
         assert_eq!(
-            builder.resolved_deployment_group_url(&project, "mr/42"),
+            builder
+                .url_builder
+                .resolved_deployment_group_url(&project, "mr/42"),
             Some("demo-mr--42.preview.example.test".to_string())
         );
     }
@@ -2551,7 +2215,7 @@ mod tests {
     #[test]
     fn primary_hosts_default_group_prod_env_includes_production_and_custom() {
         let mut builder = test_resource_builder();
-        builder.staging_ingress_url_template =
+        builder.url_builder.staging_ingress_url_template =
             Some("{project_name}-{deployment_group}.preview.example.test".to_string());
         let project = test_project();
         let prod_env = test_environment("production", true, "default");
@@ -2575,9 +2239,9 @@ mod tests {
     #[test]
     fn primary_hosts_non_default_group_with_staging_env() {
         let mut builder = test_resource_builder();
-        builder.staging_ingress_url_template =
+        builder.url_builder.staging_ingress_url_template =
             Some("{project_name}-{deployment_group}.preview.example.test".to_string());
-        builder.environment_ingress_url_template =
+        builder.url_builder.environment_ingress_url_template =
             Some("{environment}.{project_name}.example.test".to_string());
         let project = test_project();
         let staging_env = test_environment("staging", false, "mr-42");
@@ -2612,7 +2276,7 @@ mod tests {
     #[test]
     fn primary_hosts_no_env_returns_only_dg_host() {
         let mut builder = test_resource_builder();
-        builder.staging_ingress_url_template =
+        builder.url_builder.staging_ingress_url_template =
             Some("{project_name}-{deployment_group}.preview.example.test".to_string());
         let project = test_project();
 
@@ -2628,9 +2292,9 @@ mod tests {
         // collides with the env URL of the "staging" environment whose
         // primary deployment group is "staging-grp".
         let mut builder = test_resource_builder();
-        builder.staging_ingress_url_template =
+        builder.url_builder.staging_ingress_url_template =
             Some("{deployment_group}--{project_name}.preview.example.test".to_string());
-        builder.environment_ingress_url_template =
+        builder.url_builder.environment_ingress_url_template =
             Some("{environment}--{project_name}.preview.example.test".to_string());
         let project = test_project();
         let staging_env = test_environment("staging", false, "staging-grp");
@@ -2673,7 +2337,7 @@ mod tests {
     #[test]
     fn create_primary_ingress_emits_multiple_rules_and_tls_entries() {
         let mut builder = test_resource_builder();
-        builder.staging_ingress_url_template =
+        builder.url_builder.staging_ingress_url_template =
             Some("{project_name}-{deployment_group}.preview.example.test".to_string());
         builder.ingress_tls_secret_name = Some("shared-wildcard".to_string());
         builder.access_classes.insert(
@@ -2730,9 +2394,9 @@ mod tests {
         // a different group). With no other host to fall back to, the ingress
         // is suppressed entirely.
         let mut builder = test_resource_builder();
-        builder.staging_ingress_url_template =
+        builder.url_builder.staging_ingress_url_template =
             Some("{deployment_group}--{project_name}.preview.example.test".to_string());
-        builder.environment_ingress_url_template =
+        builder.url_builder.environment_ingress_url_template =
             Some("{environment}--{project_name}.preview.example.test".to_string());
         builder.access_classes.insert(
             "default".to_string(),
